+++ src/types/Mapping.ts
@@ -11,6 +11,7 @@ export type Mapping = {
     bold: boolean;
     valueBlink: boolean;
     backgroundBlink: boolean;
+    overrideValue?: string;
   };
 };